add: post of daily tides level

+++ backend/src/main/java/com/venclima/dto/StationDTO.java
@@ -2,14 +2,14 @@ package com.venclima.dto;
 
 public class StationDTO {
 
-    private long id;
+    private int id;
     private String name;
     private double latitude;
     private double longitude;
     private String name_abbreviation;
 
-    public long getId() { return id; }
-    public void setId(long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
